Tree: consolidate all CSS
