--- client/web/src/tree/components/TreeLayerCell.module.scss
@@ -1,5 +1,0 @@
-.cell {
-    padding-left: 0;
-    padding-right: 0;
-    border-radius: var(--border-radius);
-}
