--- client/web/src/tree/components/TreeLayerRowContents.module.scss
@@ -1,12 +1,0 @@
-.tree-row-contents {
-    display: flex;
-    color: inherit;
-    align-items: center;
-    padding-top: 0.25rem;
-    padding-bottom: 0.25rem;
-    cursor: pointer;
-}
-
-.is-new {
-    justify-content: space-between;
-}
