--- client/web/src/tree/components/TreeLayerRowContentsText.module.scss
@@ -1,5 +1,0 @@
-.tree-row-contents-text {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
