--- client/web/src/tree/components/TreeLayerTable.module.scss
@@ -1,4 +1,0 @@
-.tree-layer {
-    width: 100%;
-    outline: none;
-}
