--- client/web/src/tree/components/TreeRow.module.scss
@@ -1,38 +1,0 @@
-.row {
-    color: var(--body-color);
-    width: 100%;
-
-    &:hover {
-        background-color: var(--color-bg-2);
-    }
-
-    &--selected {
-        background-color: var(--color-bg-3);
-        cursor: pointer;
-
-        &:hover {
-            background-color: var(--color-bg-3);
-        }
-    }
-
-    &--active {
-        width: 100%;
-        background-color: var(--color-bg-3);
-        color: var(--body-color);
-    }
-
-    &-contents,
-    &-label {
-        &:hover {
-            text-decoration: none;
-        }
-    }
-
-    :global(.theme-light) &--active,
-    :global(.theme-light) &--selected,
-    :global(.theme-light) &:hover {
-        meter {
-            --meter-background: var(--gray-05);
-        }
-    }
-}
