--- client/web/src/tree/components/TreeRowAlert.module.scss
@@ -1,6 +1,0 @@
-.row-alert {
-    display: block;
-    padding: 0.75rem 1.25rem;
-    white-space: pre-wrap;
-    margin-bottom: 0;
-}
