--- client/web/src/tree/components/TreeRowIcon.module.scss
@@ -1,5 +1,0 @@
-.row-icon {
-    color: var(--link-color);
-    /* stylelint-disable-next-line declaration-property-unit-allowed-list */
-    width: 1.1em;
-}
