--- client/web/src/tree/components/TreeRowLabel.module.scss
@@ -1,4 +1,0 @@
-.row-label {
-    margin-left: 0.25rem;
-    color: inherit;
-}
